feat: add router and refactoring folder
