+++ test-vue/src/main.ts
@@ -4,5 +4,11 @@ import { createApp } from 'vue'
 import App from './App.vue'
 import { key, store } from './store'
 import PrimeVue from 'primevue/config'
+import router from './router'
 
-createApp(App).use(store, key).use(PrimeVue).mount('#app')
+
+createApp(App)
+    .use(router)
+    .use(store, key)
+    .use(PrimeVue)
+    .mount('#app')
