+++ test-vue/src/router/index.ts
@@ -1,0 +1,22 @@
+import { createRouter, createWebHistory } from 'vue-router'
+import HomeView from '../views/HomePage.vue'
+
+const routes = [
+  {
+    path: '/',
+    name: 'home',
+    component: HomeView
+  },
+  {
+    path: '/cart',
+    name: 'cart',
+    component: () => import('../views/CartPage.vue')
+  }
+]
+
+const router = createRouter({
+  history: createWebHistory(),
+  routes
+})
+
+export default router
