+++ test-vue/src/views/CartPage.vue
@@ -1,0 +1,12 @@
+<script setup lang="ts">
+
+console.log('12ew')
+</script>
+
+<template>
+ <div>Page Item</div>
+</template>
+
+<style scoped>
+
+</style>
